+++ db/app/ddl/src/incr/V20220716105300__drop_t_ref_constraint_object.sql
@@ -1,0 +1,1 @@
+drop type T_REF_CONSTRAINT_OBJECT force;